reorganize ingredient queries
